disarray/requirements.txt
disnake==2.4.0
Pillow==9.0.1
python-dotenv==0.19.2
aiohttp==3.8.1
deta==1.1.0
